execution/enrichment/tech_stack_detector.py
"""
Detect technology stack from website.
"""
import sys
import json
import requests
from bs4 import BeautifulSoup

def detect_tech_stack(url: str) -> dict:
    """
    Detect technologies used on a website.
    
    Args:
        url: Website URL to analyze
    
    Returns:
        Dict with detected technologies
    """
    # Placeholder implementation
    # In production, use Wappalyzer or BuiltWith API
    
    return {
        'url': url,
        'technologies': [],
        'frameworks': [],
        'analytics': [],
        'hosting': []
    }

if __name__ == "__main__":
    print("tech_stack_detector.py - Technology detection tool")
