requirements.txt
# Python dependencies for agentic workflow system

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0

# Web scraping
playwright>=1.40.0
selenium>=4.15.0

# Data validation
dnspython>=2.4.0
email-validator>=2.1.0

# Database and storage
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
supabase>=2.0.0

# AI/ML (for email generation)
openai>=1.3.0
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and monitoring
loguru>=0.7.0

# Optional: Advanced features
# wappalyzer>=0.3.0  # Tech stack detection
# linkedin-api>=2.0.0  # LinkedIn integration
